requirements.txt
# pip install "fastapi[standard]"
fastapi[standard]
logging
sqlalchemy

uvicorn[standard]
psycopg2-binary
python-nanoid
python-multipart
